# feat: implement enhanced git hooks with comprehensive validation and performance optimization