--- a/.github/workflows/kmp-ci.yml
+++ b/.github/workflows/kmp-ci.yml
@@ -2,7 +2,7 @@ name: KMP CI/CD Pipeline
 
 on:
   push:
-    branches: [ main, develop ]
+    branches: [ main ]
   pull_request:
     branches: [ main ]
   schedule: