--- a/.github/workflows/sprint-metrics.yml
+++ b/.github/workflows/sprint-metrics.yml
@@ -206,7 +206,7 @@ jobs:
             const burndown = [];
             const totalDays = 14;
 
-            for (let day = 0; day <= totalDays; day++) {
+            for (let day = 0; day < totalDays; day++) {
               const date = new Date(startDate);
               date.setDate(date.getDate() + day);