--- a/docs/COMMUNICATION.md
+++ b/docs/COMMUNICATION.md
@@ -340,4 +340,4 @@ Maintain in private repo:
 ---
 
 *Last Updated: September 27, 2025*
-*Next Review: Q1 2025*
+*Next Review: Q1 2026*