--- a/test-file.txt
+++ b/test-file.txt
@@ -1,0 +1,1 @@
+# Testing hooks