--- a/test-file.txt
+++ b/test-file.txt
@@ -1,1 +1,0 @@
-# Testing hooks