… .github/DISCUSSION_TEMPLATE/architecture-decision.yml
… .github/DISCUSSION_TEMPLATE/rfc.yml
… .github/DISCUSSION_TEMPLATE/sprint-retrospective.yml
… .github/workflows/release.yml
… CONTRIBUTING.md
… docs/TEAM_CHARTER.md